requirements.txt
altgraph==0.17.3
asgiref==3.7.2
auto-py-to-exe==2.36.0
Babel==2.12.1
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
customtkinter==5.2.0
darkdetect==0.8.0
distlib==0.3.7
Django==4.2.4
django-simple-history==3.3.0
djangorestframework==3.14.0
drf-yasg==1.21.7
Eel==0.16.0
filelock==3.12.2
future==0.18.3
gevent==23.7.0
gevent-websocket==0.10.1
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
packaging==23.1
pefile==2023.2.7
Pillow==10.0.0
platformdirs==3.10.0
protobuf==4.21.12
psycopg2==2.9.6
pycparser==2.21
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
pyparsing==3.1.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pytz==2023.3
pywin32-ctypes==0.2.2
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tkcalendar==1.6.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
virtualenv==20.24.2
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.0
